split up classes and put them into View folder
